docs/requirements.txt
qdrant-client>=1.1.1
PyGithub>=1.59.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
